src/requirements.txt
langchain-core>=0.3
langchain-anthropic>=0.3
langgraph>=0.4
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
python-pptx>=0.6.21
